+++ extensions/languages/en.php
@@ -1,11 +1,11 @@
 <?php
 
 return [
-    'beebmx.kirby-patrol.invert' => 'Invert',
-    'beebmx.kirby-patrol.invert-selection' => 'Invert selection',
-    'beebmx.kirby-patrol.name' => 'Patrol',
-    'beebmx.kirby-patrol.permission' => 'Permission',
-    'beebmx.kirby-patrol.permissions' => 'Permissions',
-    'beebmx.kirby-patrol.roles' => 'Roles',
-    'beebmx.kirby-patrol.select-all' => 'Select all',
+    'beebmx.patrol.invert' => 'Invert',
+    'beebmx.patrol.invert-selection' => 'Invert selection',
+    'beebmx.patrol.name' => 'Patrol',
+    'beebmx.patrol.permission' => 'Permission',
+    'beebmx.patrol.permissions' => 'Permissions',
+    'beebmx.patrol.roles' => 'Roles',
+    'beebmx.patrol.select-all' => 'Select all',
 ];