+++ extensions/languages/es_419.php
@@ -1,11 +1,11 @@
 <?php
 
 return [
-    'beebmx.kirby-patrol.invert' => 'Invertir',
-    'beebmx.kirby-patrol.invert-selection' => 'Invertir selección',
-    'beebmx.kirby-patrol.name' => 'Patrulla',
-    'beebmx.kirby-patrol.permission' => 'Permiso',
-    'beebmx.kirby-patrol.permissions' => 'Permisos',
-    'beebmx.kirby-patrol.roles' => 'Roles',
-    'beebmx.kirby-patrol.select-all' => 'Seleccionar todo',
+    'beebmx.patrol.invert' => 'Invertir',
+    'beebmx.patrol.invert-selection' => 'Invertir selección',
+    'beebmx.patrol.name' => 'Patrulla',
+    'beebmx.patrol.permission' => 'Permiso',
+    'beebmx.patrol.permissions' => 'Permisos',
+    'beebmx.patrol.roles' => 'Roles',
+    'beebmx.patrol.select-all' => 'Seleccionar todo',
 ];