+++ resources/index.js
@@ -2,7 +2,7 @@ import KirbyPatrolView4 from './components/KirbyPatrolView4.vue'
 import KirbyPatrolView5 from './components/KirbyPatrolView5.vue'
 import KirbyPatrolView from './components/KirbyPatrolView.vue'
 
-window.panel.plugin('beebmx/kirby-patrol', {
+window.panel.plugin('beebmx/patrol', {
   components: {
     'k-patrol-view': KirbyPatrolView,
     'k-patrol-view-4': KirbyPatrolView4,