+++ tests/Unit/ContentTest.php
@@ -51,7 +51,7 @@ describe('basic', function () {
 describe('advance', function () {
     it('returns a specific query collection', function () {
         $content = new Content(App(options: [
-            'beebmx.kirby-patrol.content.query' => function (Site $site, Pages $pages, Kirby $kirby): Pages {
+            'beebmx.patrol.content.query' => function (Site $site, Pages $pages, Kirby $kirby): Pages {
                 return $site->find('content')->children()->listed();
             },
         ]));
@@ -67,7 +67,7 @@ describe('advance', function () {
 
     it('can update the depth limit', function () {
         $content = new Content(App(options: [
-            'beebmx.kirby-patrol.content.depth' => 3,
+            'beebmx.patrol.content.depth' => 3,
         ]));
 
         expect($content->toArray())
@@ -90,9 +90,9 @@ describe('advance', function () {
 
     it('can update the sort value', function () {
         $content = new Content(App(options: [
-            'beebmx.kirby-patrol.content.depth' => 2,
-            'beebmx.kirby-patrol.content.direction' => 'desc',
-            'beebmx.kirby-patrol.content.sort' => 'num',
+            'beebmx.patrol.content.depth' => 2,
+            'beebmx.patrol.content.direction' => 'desc',
+            'beebmx.patrol.content.sort' => 'num',
         ]));
 
         expect($content->toArray())
@@ -105,7 +105,7 @@ describe('advance', function () {
 describe('invalid data', function () {
     it('returns pages even with invalid data', function () {
         $content = new Content(App(options: [
-            'beebmx.kirby-patrol.content.query' => function (Site $site, Pages $pages, Kirby $kirby) {
+            'beebmx.patrol.content.query' => function (Site $site, Pages $pages, Kirby $kirby) {
                 return $site->find('invalid');
             },
         ]));